Extensive renaming
"Piece" -> "Block"
"Block" -> "Piece"
Done in order to prevent confusion. In most BitTorrent documents, a
"Piece" is referred to as a large chunk of data with a checksum
available in the metadata file.
In contrast, a "Block" is a smaller chunk of data which peers exchange
by "Piece" messages. A block is 16KB in this implementation.